refactor(cli): split mcp_manager.go under the 800-line ceiling
The /mcp manager landed in one 1220-line file. Split along the
render/action/state seams into mcp_manager.go (state, key handling,
snapshot), mcp_manager_actions.go (connect/disable/remove/mode/auth/edit),
and mcp_manager_view.go (rendering + display strings). Same package, no
behaviour change.
